+++ lib/vtls/rustls.c
@@ -1214,8 +1214,8 @@ cr_connect(struct Curl_cfilter *cf,
         while(rustls_connection_get_peer_certificate(rconn, num_certs)) {
           num_certs++;
           if(num_certs > MAX_ALLOWED_CERT_AMOUNT) {
-            failf(data, "%d certificates is more than allowed (%u)",
-                  (int)num_certs, MAX_ALLOWED_CERT_AMOUNT);
+            failf(data, "%zu certificates is more than allowed (%u)",
+                  num_certs, MAX_ALLOWED_CERT_AMOUNT);
             return CURLE_SSL_CONNECT_ERROR;
           }
         }
